requirements.txt
python-telegram-bot[asyncio]==21.11.1
# fijamos httpx 0.27 para que ambas libs usen la misma versión
httpx==0.27.*

# SDK Solana con restricción relajada
solana>=0.33,<0.35          # 0.34.x funciona bien

# AnchorPy ya es compatible con solana 0.33+
anchorpy>=0.20,<0.21

aiohttp>=3.8
python-dotenv>=1.0
tinydb>=4.7
cryptography>=42.0
base58>=2.1
aiohttp>=3.8
Brotli

